refactor(insights-ui): remove unused public-equities pages and orphaned code
The legacy public-equities page tree is no longer a navigable feature — it's
disallowed in robots.ts, absent from all sitemaps, has no internal links, and
shows no traffic in GSC over the last 3 months.

- Relocate SpiderChartFlyoutMenu (used by live stock + ETF report pages) from
  the public-equities tree to src/components/ticker/, updating the 3 importers.
- Delete src/app/public-equities/** (31 files).
- Delete now-orphaned components used only by that tree: criteria/CriteriaTable,
  criteria/UpsertCustomCriteria, ticker/EditTickerView, and the four
  ticker/debug/* components (DebugManagementTeam, DebugMatchingAttachments,
  DebugFinancialStatements, ManagementDiscussionButton).
- Delete the now-orphaned api/custom-criteria route (only caller was CriteriaTable).
- Drop /public-equities from robots.ts disallow and the login popup exclusion list.

Shared backend (industries, sub-industries, industry-analysis, analysis-*,
actions, tickers) is intentionally left intact — it's still used by live
admin-v1, stocks, prompts, and portfolio-managers pages.
