--- insights-ui/src/app/public-equities/debug/tickers/[ticker]/criteria/[criterion]/page.tsx
@@ -1,12 +1,0 @@
-import PageWrapper from '@dodao/web-core/components/core/page/PageWrapper';
-import CriterionDebugPage from './CriterionDebugPage';
-
-export default async function Page({ params }: { params: Promise<{ ticker: string; criterion: string }> }) {
-  const { ticker, criterion } = await params;
-
-  return (
-    <PageWrapper>
-      <CriterionDebugPage ticker={ticker} criterionKey={criterion} />
-    </PageWrapper>
-  );
-}
