--- insights-ui/src/app/public-equities/tickers/[tickerKey]/criteria/[criterionKey]/page.tsx
@@ -1,42 +1,0 @@
-import { permanentRedirect, notFound } from 'next/navigation';
-import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
-import { KoalaGainsSpaceId } from '@/types/koalaGainsConstants';
-
-export default async function CriterionDetailsPage({ params }: { params: Promise<{ tickerKey: string; criterionKey: string }> }) {
-  let { tickerKey, criterionKey } = await params;
-
-  // Decode the URL parameters and check if they contain '}' character
-  const decodedTickerKey = decodeURIComponent(tickerKey);
-  const decodedCriterionKey = decodeURIComponent(criterionKey);
-
-  if (decodedTickerKey.includes('}') || decodedCriterionKey.includes('}')) {
-    const cleanedTickerKey = decodedTickerKey.replace(/\}/g, '');
-    tickerKey = cleanedTickerKey;
-  }
-
-  // Fetch exchange information for the ticker to redirect to new URL format
-  try {
-    const exchangeResponse = await fetch(`${getBaseUrl()}/api/${KoalaGainsSpaceId}/tickers-v1/${tickerKey}/get-exchange-name`, { cache: 'no-cache' });
-
-    if (exchangeResponse.ok) {
-      const exchangeData = await exchangeResponse.json();
-      const exchange = exchangeData.exchange;
-
-      if (exchange) {
-        permanentRedirect(`/stocks/${exchange}/${tickerKey}`);
-      } else {
-        // Ticker not found in database
-        notFound();
-      }
-    } else {
-      // Server error (500, 503, etc.) - throw to show error page
-      throw new Error(`API returned ${exchangeResponse.status}: ${exchangeResponse.statusText}`);
-    }
-  } catch (error) {
-    if (error instanceof Error && error.message.includes('NEXT_REDIRECT')) {
-      throw error;
-    }
-    // Network error or other server issue - throw to show error page
-    throw error;
-  }
-}
