+++ insights-ui/src/app/robots.ts
@@ -10,7 +10,7 @@ export default function robots(): MetadataRoute.Robots {
         // Do NOT block /_next/ — Googlebot needs to fetch /_next/static/* (CSS/JS) and
         // /_next/image (optimized images, including the logo) to render pages correctly.
         // Blocking it strips styling and breaks images in Search Console's URL Inspection.
-        disallow: ['/public-equities', '/api'],
+        disallow: ['/api'],
       },
     ],
     sitemap: 'https://koalagains.com/sitemap_index.xml',
