+++ insights-ui/src/components/etf-reportsv1/analysis/EtfRadarChartSvg.tsx
@@ -1,6 +1,6 @@
 import type { EtfScoresResponse } from '@/types/etf/etf-detail-response-types';
 import { EtfAnalysisResponse } from '@/app/api/[spaceId]/etfs-v1/exchange/[exchange]/[etf]/analysis/route';
-import SpiderChartFlyoutMenu from '@/app/public-equities/tickers/[tickerKey]/SpiderChartFlyoutMenu';
+import SpiderChartFlyoutMenu from '@/components/ticker/SpiderChartFlyoutMenu';
 import TickerRadarChartSvg from '@/components/visualizations/TickerRadarChartSvg';
 import { getSpiderGraphScorePercentage } from '@/util/radar-chart-utils';
 import { buildEtfSpiderGraph } from '@/utils/etf-analysis-reports/etf-spider-graph';
