+++ insights-ui/src/components/ticker/SpiderChartFlyoutMenu.tsx
